Convert app to new libraries and publish messages to Pulse
This is a reworking of the existing (unused) taskcluster-treeherder
application.  It will make use of the new TaskCluster libraries as well
as publish messages to Pulse for anyone to consume, including
Treeherder, instead of using the Treeherder API.

Schemas have been added that will validate the task configuration when
receiving task events from the queue, as well as validating the job
message prior to posting to the jobs exchange.